+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Add option to allow pushing on a Stack when this is full
+- Add `EMPTY` operation
 
 ### Changed
 - pkg/stack: Use RWMutex for concurrent reads